refactor: move function
Move manifest related functions from core to own module
