+++ test/test-core.ts
@@ -7,10 +7,8 @@ import {
   compareEditorVersion,
   env,
   fetchPackageInfo,
-  loadManifest,
   parseEditorVersion,
   parseEnv,
-  saveManifest,
 } from "../src/core";
 
 import {
@@ -25,6 +23,7 @@ import {
 import testConsole from "test-console";
 import assert from "assert";
 import { isInternalPackage } from "../src/utils/pkg-name";
+import { loadManifest, saveManifest } from "../src/utils/manifest";
 
 describe("cmd-core.ts", function () {
   describe("parseEnv", function () {
@@ -202,82 +201,7 @@ describe("cmd-core.ts", function () {
     });
   });
 
-  describe("loadManifest/SaveManifest", function () {
-    let stdoutInspect: testConsole.Inspector = null!;
-    let stderrInspect: testConsole.Inspector = null!;
-    beforeEach(function () {
-      removeWorkDir("test-openupm-cli");
-      createWorkDir("test-openupm-cli", { manifest: true });
-      createWorkDir("test-openupm-cli-wrong-json", {
-        manifest: true,
-      });
-      fs.writeFileSync(
-        path.join(
-          getWorkDir("test-openupm-cli-wrong-json"),
-          "Packages/manifest.json"
-        ),
-        "wrong-json"
-      );
-      [stdoutInspect, stderrInspect] = getInspects();
-    });
-    afterEach(function () {
-      removeWorkDir("test-openupm-cli");
-      removeWorkDir("test-openupm-cli-wrong-json");
-      stdoutInspect.restore();
-      stderrInspect.restore();
-    });
-    it("loadManifest", async function () {
-      (
-        await parseEnv(
-          { _global: { chdir: getWorkDir("test-openupm-cli") } },
-          { checkPath: true }
-        )
-      ).should.be.ok();
-      const manifest = loadManifest();
-      assert(manifest !== null);
-      manifest.should.be.deepEqual({ dependencies: {} });
-    });
-    it("no manifest file", async function () {
-      (
-        await parseEnv(
-          { _global: { chdir: getWorkDir("path-not-exist") } },
-          { checkPath: false }
-        )
-      ).should.be.ok();
-      const manifest = loadManifest();
-      (manifest === null).should.be.ok();
-      const [stdout] = getOutputs(stdoutInspect, stderrInspect);
-      stdout.includes("does not exist").should.be.ok();
-    });
-    it("wrong json content", async function () {
-      (
-        await parseEnv(
-          { _global: { chdir: getWorkDir("test-openupm-cli-wrong-json") } },
-          { checkPath: true }
-        )
-      ).should.be.ok();
-      const manifest = loadManifest();
-      (manifest === null).should.be.ok();
-      const [stdout] = getOutputs(stdoutInspect, stderrInspect);
-      stdout.includes("failed to parse").should.be.ok();
-    });
-    it("saveManifest", async function () {
-      (
-        await parseEnv(
-          { _global: { chdir: getWorkDir("test-openupm-cli") } },
-          { checkPath: true }
-        )
-      ).should.be.ok();
-      const manifest = loadManifest();
-      assert(manifest !== null);
-      manifest.should.be.deepEqual({ dependencies: {} });
-      manifest.dependencies["some-pack"] = "1.0.0";
-      saveManifest(manifest).should.be.ok();
-      const manifest2 = loadManifest();
-      assert(manifest2 !== null);
-      manifest2.should.be.deepEqual(manifest);
-    });
-  });
+  
 
   describe("fetchPackageInfo", function () {
     beforeEach(function () {
